requirements.txt
tiktoken==0.9.0
autogen-core==0.4.7
autogen-agentchat==0.4.7
autogen-ext[azure]==0.4.7
autogen-ext[openai]==0.4.7
autogen-ext[docker]==0.4.7
SQLAlchemy==2.0.37
aiohttp==3.11.11
aiosqlite==0.20.0
chainlit==2.0.603
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.10.1
rich==13.9.4
azure-ai-projects==1.0.0b6
azure-ai-evaluation==1.3.0
azure-ai-ml==1.25.0
azure-identity==1.20.0
promptflow-devkit==1.17.2
azure-search-documents==11.5.2